requirements.txt
#
####### telegram-requirements.txt #######
#
###### Requirements without Version Specifiers ######
mtgsdk
pillow
bs4
datetime
requests
matplotlib
numpy
robobrowser
unidecode
reportlab
requests_oauthlib
#
###### Requirements with Version Specifiers ######
#   See https://www.python.org/dev/peps/pep-0440/#version-specifiers
python-telegram-bot >= 8.0
#